+++ components/navigation.php
@@ -128,7 +128,19 @@ $tools_items = [
         'icon' => 'fas fa-plug',
         'url' => '?page=integration-tools',
         'description' => 'Integration tools'
-    ]
+    ],
+    'journal-archive' => [
+        'title' => 'Journal Archive',
+        'icon' => 'fas fa-book-open',
+        'url' => '?page=journal-archive',
+        'description' => 'Arsip jurnal Scopus'
+    ],
+    'sdg-researcher-list' => [
+        'title' => 'SDG Researchers',
+        'icon' => 'fas fa-users',
+        'url' => '?page=sdg-researcher-list',
+        'description' => 'Daftar peneliti per SDG'
+    ],
 ];
 ?>